+++ src/Admin/Controls/ProductGridFiltersFactory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Eshop\Admin\Controls;
 
 use Admin\Controls\AdminGrid;
-use Eshop\Common\Helpers;
+use Eshop\Common\Services;
 use Eshop\DB\CategoryRepository;
 use Eshop\DB\CategoryTypeRepository;
 use Eshop\DB\DisplayAmountRepository;